group order fills by id and show one mark per order
