+++ src/hooks/tradingView/useTradingView.ts
@@ -147,7 +147,7 @@ export const useTradingView = ({
             }
             if (buySellMarksToggleRef) {
               buySellMarksToggleRef.current = tvWidgetRef.current.createButton();
-              buySellMarksToggleRef.current.innerHTML = `<span>Buys/Sells</span> <div class="toggle"></div>`;
+              buySellMarksToggleRef.current.innerHTML = `<span>Buys/Sells</span> <div class="displayOrdersButton-toggle"></div>`;
               buySellMarksToggleRef.current.setAttribute(
                 'title',
                 'Display historic buys and sells'
